♻️ Move test to integration-tests file
